feat: Add simple hashmap implementation
This hashmap maps 2 points (4-int coordinates) to an int.
Its intended use was to cache results of euclidean distance between 2
points, because I profiled the code and that's the bottleneck currently.
But I scrapped the idea, because it's actually much slower than the
euclidean distance function. Still, I might find it useful for something
eventually.