feat(graph): Add distance matrix caching
This was actually one of the task requirements that I did not see.
And this optimization method actually improves times majorly, unlike
hashmap.

I HAD thought about it as a better alternative to hashmap (8fbd3b7557),
but I deemed it unfeasible due to possible excess memory usage (without
any assumptions about graph size). However, since the task requires it,
I suppose memory usage is guaranteed to be sufficiently low.

NOTE: Some deterministic solutions have changed, even though they
shouldn't. This suggests the distances are somehow different than prior
to this commit. I'm confident the code is safe and there is no undefined
behavior, this is probably a result of rounding. I don't have the time
to look for the exact reason, the results are similar anyway.