fix(graph)!: Use node cost information for nn, 2nn, nc
The former behavior was intentional on my part, but it turned out that
it was incorrect with regards to the tasks, so I am correcting it now.

BREAKING CHANGE: Naturally, this drastically improves results in task1,
but I am not updating them to keep the report faithful to what I turned
in. The updated scores will be made available in task2.