feat(graph): Add tsp_graph_find_rcl()
Meant for task2, this new function does the first annoying step of
choosing a sample of N best node candidates for activation in a graph.

Also added a generic shuffle() function that's handy for shuffling
buffers in-place. I use it in tsp_graph_find_rcl() to avoid copying
memory back and forth.